refactor: extract DrawingML color primitives into shared module
DrawingML color markup (srgbClr/schemeClr/sysClr with nested tint,
shade, lumMod, lumOff, alpha children) is the same across pptx, docx,
and xlsx parts, but only pptx_theme.rs implemented the full transform
math. Move scheme-color resolution, transform parsing/application, the
color-element readers, and the rgb/hsl conversions verbatim into a new
format-agnostic parser/drawingml.rs behind a SchemeColors view (theme
palette + optional alias map).

pptx_theme.rs keeps its signatures as thin delegating wrappers, so the
~20 pptx call sites are untouched and output is byte-identical (full
suite passes unchanged; 10 new unit tests cover the shared module).

Groundwork for routing docx and xlsx DrawingML colors — currently a
weaker fork without transforms and a hardcoded black/white schemeClr
fallback respectively — through the same math in follow-up changes.
