+++ crates/office2pdf/src/parser/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod chart;
 pub(crate) mod cond_fmt;
 pub mod docx;
+pub(crate) mod drawingml;
 pub(crate) mod embedded_fonts;
 #[path = "pptx_emf.rs"]
 pub(crate) mod emf;
